Namespace to snapshot
